codec multi value alias reorg
