+++ elrond-wasm/src/types/io/mod.rs
@@ -1,20 +1,14 @@
 mod async_call_result;
-mod ignore_args;
-mod multi_args;
-mod multi_args_vec;
+mod codec_multi_value_aliases;
 mod operation_completion_status;
-mod optional_arg;
 mod sc_error;
 mod sc_error_managed;
 mod sc_error_static;
 mod sc_result;
 
 pub use async_call_result::{AsyncCallError, AsyncCallResult};
-pub use ignore_args::IgnoreVarArgs;
-pub use multi_args::*;
-pub use multi_args_vec::{MultiArgVec, MultiResultVec, VarArgs};
+pub use codec_multi_value_aliases::*;
 pub use operation_completion_status::OperationCompletionStatus;
-pub use optional_arg::{OptionalArg, OptionalResult};
 pub use sc_error::SCError;
 pub use sc_error_managed::ManagedSCError;
 pub use sc_error_static::StaticSCError;
